requirements.txt
fuzzywuzzy==0.10.0
python-Levenshtein==0.12.1
PyYAML==5.4
requests==2.23.0
# To install discord do pip install -U discord.py
